requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
drf-extensions==0.7.0
httpie==2.4.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy==0.812
mypy-extensions==0.4.3
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
